:recycle:주석 및 쓰지 않는 함수 삭제.
